Refactor to remediate aforementioned tangle of code
Moves the logic to figure out the file(s) to use inside the Simulation objects. All that the plugin cares about is whether the file(s) specified
exist, not how they got there.
